Move, update, and fix futures-channel tests
