+++ .travis.yml
@@ -17,15 +17,20 @@ matrix:
         - travis-cargo doc-upload
     - rust: stable
       script:
-        - cargo test -p futures-core --no-default-features
-        - cargo test -p futures --no-default-features
-        - cargo test -p futures-channel --no-default-features
-        - cargo test -p futures-cpupool --no-default-features
-        - cargo test -p futures-executor --no-default-features
-        - cargo test -p futures-sink --no-default-features
-        - cargo test -p futures-util --no-default-features
+        - cargo build -p futures-core --no-default-features
+        - cargo build -p futures --no-default-features
+        - cargo build -p futures-channel --no-default-features
+        - cargo build -p futures-cpupool --no-default-features
+        - cargo build -p futures-executor --no-default-features
+        - cargo build -p futures-sink --no-default-features
+        - cargo build -p futures-util --no-default-features
     - rust: 1.15.0
       script: cargo test
+
+script:
+  - cargo test --all
+  - cargo test --all --release
+
 env:
   global:
     - secure: "iwVcMVIF7ZSY82fK5UyyUvVvJxMSYrbZawh1+4Oi8pvOdYq1gptcDoOC8jxWwCwrNF1b+/85n+jlEUngEqqSmV5PjAbWPjoc+u4Zn7CRi1AlxoUlvHPiQm4vM4Mkkd6GsqoIZttCeedU9m/w0nQ18uUtK8uD6vr2FVdcMnUnkYQAxuGOowGLrwidukzfBXMCu/JrwKMIbt61knAFiI/KJknu0h1mRrhpeF/sQ3tJFzRRcQeFJkbfwDzltMpPo1hq5D3HI4ONjYi/qO2pwUhDk4umfp9cLW9MS8rQvptxJTQmWemHi+f2/U4ld6a0URL6kEuMkt/EbH0A74eFtlicfRs44dX9MlWoqbLypnC3ymqmHcpwcwNA3HmZyg800MTuU+BPK41HIPdO9tPpxjHEiqvNDknH7qs+YBnis0eH7DHJgEjXq651PjW7pm+rnHPwsj+OzKE1YBNxBQZZDkS3VnZJz+O4tVsOzc3IOz0e+lf7VVuI17C9haj117nKp3umC4MVBA0S8RfreFgqpyDeY2zwcqOr0YOlEGGRl0vyWP8Qcxx12kQ7+doLolt6Kxda4uO0hKRmIF6+qki1T+L7v8BOGOtCncz4f7IX48eQ7+Wu0OtglRn45qAa3CxjUuW6xX3KSNH66PCXV0Jtp8Ga2SSevX2wtbbFu9f+9R+PQY4="
