+++ futures-channel/Cargo.toml
@@ -19,3 +19,7 @@ version = "0.1.0"
 path = "../futures-core"
 default-features = false
 features = ["std"]
+
+[dev-dependencies]
+futures = { path = "../futures", version = "0.2" }
+futures-executor = { path = "../futures-executor", version = "0.1" }
