+++ futures/src/lib.rs
@@ -177,8 +177,8 @@ pub use futures_util::sink::SinkExt;
 #[macro_export]
 macro_rules! try_ready {
     ($e:expr) => (match $e {
-        Ok($crate::Async::Ready(t)) => t,
-        Ok($crate::Async::Pending) => return Ok($crate::Async::Pending),
+        Ok($crate::prelude::Async::Ready(t)) => t,
+        Ok($crate::prelude::Async::Pending) => return Ok($crate::prelude::Async::Pending),
         Err(e) => return Err(From::from(e)),
     })
 }
@@ -280,7 +280,7 @@ pub mod stream {
 	//! https://tokio.rs][online]
 	//!
 	//! [online]: https://tokio.rs/docs/getting-started/streams-and-sinks/
-    
+
     pub use futures_core::stream::*;
     pub use futures_util::stream::*;
 }
